requirements.txt
amqp==5.2.0
APScheduler==3.10.4
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
cachetools==5.4.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.40.0
cryptography==43.0.0
defusedxml==0.8.0rc2
dj-database-url==2.2.0
dj-rest-auth==6.0.0
Django==5.0.7
django-allauth==64.0.0
django-auth-adfs==1.14.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
django-crontab==0.7.1
django-oauth-toolkit==2.4.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djangosaml==1.0.6
djangosaml2==1.9.3
elementpath==4.4.0
frozendict==2.3.10
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
gunicorn==22.0.0
idna==3.7
jwcrypto==1.5.6
jwt==1.3.1
kombu==5.3.7
msal==1.30.0
oauthlib==3.2.2
onelogin==3.1.6
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
pysaml2==7.5.0
python-dateutil==2.7.5
python-dotenv==1.0.1
python-magic==0.4.27
python3-openid==3.2.0
pytz==2024.1
redis==5.0.8
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sentry-sdk==2.12.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.19
vine==5.1.0
wcwidth==0.2.13
xmlschema==2.5.1
